+++ src/lib/components/action-msg/MsgToken.tsx
@@ -1,41 +1,34 @@
 import { Flex, Text } from "@chakra-ui/react";
-import type { Coin } from "@cosmjs/stargate";
+import { isUndefined } from "lodash";
 
 import { Copier } from "../copy";
 import { TooltipInfo } from "../Tooltip";
-import { formatBalanceWithDenom } from "lib/utils";
+import type { TokenWithValue } from "lib/types";
+import { formatTokenWithValue } from "lib/utils";
 
 interface MsgTokenProps {
-  coin: Coin;
-  symbol?: string;
-  precision?: number;
+  token: TokenWithValue;
   fontWeight?: number;
   ampCopierSection?: string;
 }
 
 export const MsgToken = ({
-  coin,
-  symbol,
-  precision,
+  token,
   fontWeight = 600,
   ampCopierSection,
 }: MsgTokenProps) => (
   <Flex role="group" align="center" gap={1}>
     <Text fontWeight={fontWeight} variant="body2">
-      {formatBalanceWithDenom({
-        coin,
-        symbol,
-        precision,
-      })}
+      {formatTokenWithValue(token)}
     </Text>
     <TooltipInfo
-      label={`Token ID: ${coin.denom}`}
+      label={`Token ID: ${token.denom}`}
       maxW="240px"
       textAlign="center"
     />
     <Copier
-      type={symbol ? "supported_asset" : "unsupported_asset"}
-      value={coin.denom}
+      type={isUndefined(token.price) ? "supported_asset" : "unsupported_asset"}
+      value={token.denom}
       copyLabel="Token ID Copied!"
       display="none"
       ml={1}