+++ src/lib/pages/pools/components/pool-details/tables/pool-txs/messages/swap/MsgSwapExactAmountOutAction.tsx
@@ -2,12 +2,13 @@ import { Flex } from "@chakra-ui/react";
 
 import { MsgToken } from "lib/components/action-msg/MsgToken";
 import { CustomIcon } from "lib/components/icon";
-import type { AssetInfosOpt } from "lib/services/assetService";
+import type { AssetInfos, Option } from "lib/types";
+import { coinToTokenWithValue } from "lib/utils";
 import type { MsgSwapExactAmountOutDetails } from "lib/utils/tx/types";
 
 interface MsgSwapExactAmountOutActionProps {
   msg: MsgSwapExactAmountOutDetails;
-  assetInfos: AssetInfosOpt;
+  assetInfos: Option<AssetInfos>;
   ampCopierSection?: string;
 }
 
@@ -17,23 +18,27 @@ export const MsgSwapExactAmountOutAction = ({
   ampCopierSection,
 }: MsgSwapExactAmountOutActionProps) => {
   const tokenInDenom = msg.routes[0]?.tokenInDenom ?? "";
-  const inAssetInfo = assetInfos?.[tokenInDenom];
-  const outAssetInfo = assetInfos?.[msg.token_out.denom];
+  const inToken = coinToTokenWithValue(
+    tokenInDenom,
+    msg.token_in_max_amount,
+    assetInfos
+  );
+  const outToken = coinToTokenWithValue(
+    msg.token_out.denom,
+    msg.token_out.amount,
+    assetInfos
+  );
   return (
     <Flex gap={1} alignItems="center" flexWrap="wrap">
       Swap at most
       <MsgToken
-        coin={{ amount: msg.token_in_max_amount, denom: tokenInDenom }}
-        symbol={inAssetInfo?.symbol}
-        precision={inAssetInfo?.precision}
+        token={inToken}
         fontWeight={400}
         ampCopierSection={ampCopierSection}
       />
       <CustomIcon name="arrow-right" boxSize={4} color="accent.main" />
       <MsgToken
-        coin={msg.token_out}
-        symbol={outAssetInfo?.symbol}
-        precision={outAssetInfo?.precision}
+        token={outToken}
         fontWeight={700}
         ampCopierSection={ampCopierSection}
       />