+++ CHANGELOG.md
@@ -39,9 +39,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#711](https://github.com/alleslabs/celatone-frontend/pull/711) Refactor assetInfos and add movePoolInfos to tx details
+
 ### Improvements
 
-- [#709](https://github.com/alleslabs/celatone-frontend/pull/709) refactor all address types
+- [#710](https://github.com/alleslabs/celatone-frontend/pull/710) Refactor all address types
 
 ### Bug fixes